visual_timer_gemini/requirements.txt
google-generativeai
pydub
python-dotenv
librosa
Flask
requests
streamlit
